refactor(cli): extract Intel NPU command handlers
